Fix unit tests for remote options
